Need to fix Airplane

+++ app/src/main/java/com/dolor/destroyordefense/ArenaUtilities/TypeConverter.java
@@ -21,7 +21,9 @@ class TypeConverter {
     int toPixel() {
         if (type == Type.point) {
             type = Type.pixel;
-            value *= square;
+            value *= (double) (square) ;
+
+
         }
         return value;
     }
